+++ core-ng/src/main/java/core/framework/api/util/Exceptions.java
@@ -13,7 +13,15 @@ public final class Exceptions {
         return writer.toString();
     }
 
-    public static Error error(String pattern, Object... arguments) {
+    public static Error error(String pattern, Object... arguments) {    // follow convention of logger, if last arg is exception, make it cause
+        if (arguments.length > 0) {
+            Object lastArgument = arguments[arguments.length - 1];
+            if (lastArgument instanceof Throwable) {
+                Object[] messageArguments = new Object[arguments.length - 1];
+                System.arraycopy(arguments, 0, messageArguments, 0, arguments.length - 1);
+                return new Error(Strings.format(pattern, messageArguments), (Throwable) lastArgument);
+            }
+        }
         return new Error(Strings.format(pattern, arguments));
     }
 }