+++ log-processor/src/main/java/core/log/LogProcessorApp.java
@@ -1,12 +1,12 @@
 package core.log;
 
 import core.framework.api.App;
-import core.framework.api.module.SystemModule;
 import core.log.domain.ActionDocument;
 import core.log.domain.StatDocument;
 import core.log.domain.TraceDocument;
 import core.log.job.CleanupOldIndexJob;
 import core.log.service.ActionManager;
+import core.log.service.KafkaConsumerFactory;
 import core.log.service.MessageProcessor;
 import core.log.service.StatManager;
 
@@ -18,16 +18,18 @@ import java.time.LocalTime;
 public class LogProcessorApp extends App {
     @Override
     protected void initialize() {
-        load(new SystemModule("sys.properties"));
+        loadProperties("sys.properties");
 
+        search().host(requiredProperty("sys.elasticsearch.host"));
         search().type(ActionDocument.class);
         search().type(TraceDocument.class);
         search().type(StatDocument.class);
 
-        ActionManager actionManager = bind(ActionManager.class);
-        StatManager statManager = bind(StatManager.class);
+        bind(ActionManager.class);
+        bind(StatManager.class);
 
-        MessageProcessor processor = new MessageProcessor(requiredProperty("sys.log.kafka.uri"), actionManager, statManager);
+        bind(KafkaConsumerFactory.class, new KafkaConsumerFactory(requiredProperty("sys.kafka.uri")));
+        MessageProcessor processor = bind(MessageProcessor.class);
         onStartup(processor::start);
         onShutdown(processor::stop);