+++ log-processor/src/test/java/core/log/TestModule.java
@@ -1,6 +1,10 @@
 package core.log;
 
 import core.framework.api.AbstractTestModule;
+import core.log.service.KafkaConsumerFactory;
+import org.apache.kafka.clients.consumer.Consumer;
+import org.apache.kafka.clients.consumer.MockConsumer;
+import org.apache.kafka.clients.consumer.OffsetResetStrategy;
 
 /**
  * @author neo
@@ -8,6 +12,13 @@ import core.framework.api.AbstractTestModule;
 public class TestModule extends AbstractTestModule {
     @Override
     protected void initialize() {
+        overrideBinding(KafkaConsumerFactory.class, null, new KafkaConsumerFactory("localhost:9092") {
+            @Override
+            public Consumer<String, byte[]> create() {
+                return new MockConsumer<>(OffsetResetStrategy.EARLIEST);
+            }
+        });
+
         load(new LogProcessorApp());
 
         initSearch().createIndexTemplate("action", "action-index-template.json");